split former engine.h into separate files, models.h for game board data structures, solver.h for algorithmic analysis tools on game boards
